Add as-yet unused saga-based approach to auto-saving
Hopefully this ends up simplifying the components so they don't need to
manage setTimeout handles and so that auto-save tasks are actually
cancelled.

The idea is that when you click "Close" or "Add" we'll intercept the
Link's onClick and dispatch a SAVE_EDIT_CARD with the NAVIGATE action
stuffed into the SAVE_EDIT_CARD action's onSuccess field.

This should also be more testable and a little more consistent (instead
of sometimes using Promises in some places etc.)