Mostly revert change that applied delayed focusing
Changeset b5ac5f81be9f819ec5557816f9ff7a12f65c03b9

On mobile it proved distracting because you'd close the onscreen
keyboard, do the animation for a few milliseconds, then re-open it.

If we want to do this, we should probably only do it if we don't already
have a textbox focussed (i.e. if the onscreen keyboard is not already
showing).