Switch back to draft-js
For now anyway.

Slate has a very nice API and active community, but it doesn't seem
quite ready. Most notably, there are quite a number of bugs when using
IMEs, and there are focus issues in Firefox. These are surely fixable
but it makes me wonder how many more similar bugs are still lurking
there.

Furthermore, there is a very significant package size difference, some
114Kb or so (minified)! So it means some measure of code-splitting will
be necessary right from the get go.

For the time being, draft-js seems much more battle-tested and has
about 4~5 times the number of users on npm so it's probably the safer
option for now.

I'll try to design the data format and API such that it should be
possible to switch to Draft again in future should it stabilize and get
the package size improvements it needs.