Hopefully fix problem with updating NoteList after being unmounted
When reviewing quickly we could get errors like:

```
Warning: Can't perform a React state update on an unmounted component. This is a no-op, but it indicates a memory leak in your application. To fix, cancel all subscriptions and asynchronous tasks in the componentWillUnmount method.
in NoteList (created by DynamicNoteListInner)
in DynamicNoteListInner (created by Context.Consumer)
in Connect(DynamicNoteListInner) (created by ReviewPanel) react-dom.development.js:506
```

because we'd call `setState` after unmounting.

I know you're not supposed to track mounted state but it makes a lot
more sense in this case than maintaining a list of otherwise
uninteresting Animations just so you can cancel them.